Feature/exception handling (#107)
Added support for:

- Handling any failure

- Handling specific exceptions and errors

- Handling any exception

- Handling any error

+++ server/src/main/java/io/littlehorse/common/model/getable/core/wfrun/ThreadRunModel.java
@@ -506,7 +506,7 @@ public class ThreadRunModel extends LHSerializable<ThreadRun> {
         thr.handlingFailure = new HandlingFailureHaltReasonModel();
         thr.handlingFailure.handlerThreadId = handlerThreadNumber;
 
-        childThreadIds.add((Integer) handlerThreadNumber);
+        childThreadIds.add(handlerThreadNumber);
         haltReasons.add(thr);
     }
